requirements.txt
streamlit
google-generativeai
mysql-connector-python
psycopg2-binary
pandas
plotly
sqlalchemy
langchain
urllib3
plotly-express
